Include version check in sign command
